backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-pptx==0.6.23
requests==2.32.3
